Refactored px4io_flash module to px4_flash, because it is now also used to flash the f4
